==> docs/00_data_prep.Rmd <==
---
title: "R and spatial data: data preparation"
author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    highlight: pygments
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
    toc_float: yes
---

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

library(tidyverse)
library(readxl)
library(janitor)
library(sf)
library(tmap)
```

# What is it?

Short script to prepare the dataset used in the course. Not needed by students but left for the curious ones. Simple data preps steps trying to remove some problems, making data smaller, etc.

# Inside Airbnb

Inside Airbnb data for Melbourne obtained from http://insideairbnb.com/about.html. Verision of the data (`Date compiled`) is `07 December, 2018`.

Selected variables were kept and file was resaved as `csv` without any further modifications.

```{r}
listings <- read_csv("data/raw/inside_airbnb/listings.csv") %>% 
  select(-neighbourhood_group, - minimum_nights, -number_of_reviews, -last_review, -reviews_per_month, -calculated_host_listings_count, -availability_365)

write_csv(listings, "data/listings.csv")
```

# Traffic census

Traffic census for the Queensland state-declared road network obtained from https://data.qld.gov.au/dataset/traffic-census-for-the-queensland-state-declared-road-network. 

Selected variables were kept and file was resaved as `csv` without any further modifications.

```{r}
TC2018 <- read_csv("data/raw/traffic_census/trafficcensus2018.csv")

TC2018_geo <- TC2018 %>% 
  st_as_sf(coords = c("LONGITUDE", "LATITUDE"), crs = 4326, remove = FALSE)

tmap_mode("view")

TC2018_geo %>% 
  tm_shape() + 
  tm_symbols(size = "AADT", col = "PERCENT_HV", alpha = 0.5, border.lwd = NA, id = "DESCRIPTION")

saveRDS(TC2018_geo, "data/TC2018.Rds")
```

# SEIFA

## Tabular data

Socio-Economic Indexes for Areas (SEIFA) obtained from [ABS](http://www.abs.gov.au/AUSSTATS/abs@.nsf/Lookup/2033.0.55.001Main+Features12016?OpenDocument)

Data preparation included cleaning names, removing empty rows and columns and creating more sensible variable names.

```{r}
SEIFA <- read_excel("data/raw/abs/2033055001 - sa2 indexes.xls", 
                            sheet = "Table 1", skip = 5, na = "-") %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  dplyr::rename(SA2_MAIN16 = `x1`, SA2_NAME16 = `x2`, 
                IRSD_s = `score_3`, IRSD_d = `decile_4`,
                IRSAD_s = `score_5`, IRSAD_d = `decile_6`, 
                IER_s = `score_7`, IER_d = `decile_8`, 
                IEO_s = `score_9`, IEO_d = `decile_10`, 
                URP = `x11`) %>% 
  dplyr::mutate(IRSD_d = as.integer(IRSD_d),
                IRSAD_d = as.integer(IRSAD_d),
                IER_d = as.integer(IER_d),
                IEO_d = as.integer(IEO_d)) %>% 
  select(-SA2_NAME16)

saveRDS(SEIFA, "data/SEIFA.Rds")
```


## SA2s

SA2 files from [Australian Statistical Geography Standard (ASGS)](http://www.abs.gov.au/ausstats/abs@.nsf/mf/1270.0.55.001)

Data for GCC_NAME16 Greater Melbourne selected. SA3, GCC and STE variables removed. `SA2_MAIN16` was converted from factor to numeric to ease joins.

```{r warning=FALSE}
SA2 <- st_read("./data/raw/abs/1270055001_sa2_2016_aust_shape/SA2_2016_AUST.shp") 

SA2_2016_MELB <- SA2 %>% 
  filter(GCC_NAME16 == "Greater Melbourne") %>% 
  select(-starts_with("SA3"), -starts_with("GCC"), -starts_with("STE"), -AREASQKM16, -SA2_5DIG16)

SA2_2016_MELB$SA2_MAIN16 <- as.numeric(as.character(SA2_2016_MELB$SA2_MAIN16))

st_write(SA2_2016_MELB, "./data/SA2_2016_MELB.shp")

```
