==> docs/98_extras.Rmd <==

```{r}
library(spdep)
library(ncf)
library(readr)
library(dplyr)
library(tmap)
```

# Class intervals

```{r}
knitr::opts_chunk$set(echo = FALSE)

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

library(spData)
library(classInt)

data(jenks71, package="spData")
pal1 <- c("wheat1", "red3")
opar <- par(mfrow=c(2,3))
plot(classIntervals(jenks71$jenks71, n=5, style="fixed",
                    fixedBreaks=c(15.57, 25, 50, 75, 100, 155.30)), pal=pal1, main="Fixed")
plot(classIntervals(jenks71$jenks71, n=5, style="sd"), pal=pal1, main="Pretty standard deviations")
plot(classIntervals(jenks71$jenks71, n=5, style="equal"), pal=pal1, main="Equal intervals")
plot(classIntervals(jenks71$jenks71, n=5, style="quantile"), pal=pal1, main="Quantile")
set.seed(1)
plot(classIntervals(jenks71$jenks71, n=5, style="kmeans"), pal=pal1, main="K-means")
plot(classIntervals(jenks71$jenks71, n=5, style="hclust", method="complete"),
     pal=pal1, main="Complete cluster")

```


# Lisa on point data

## `ncf`

```{r}
listings_sf_proj$X <-  st_coordinates(listings_sf_proj)[ , 1]
listings_sf_proj$Y <-  st_coordinates(listings_sf_proj)[ , 2]

# plot(listings_sf_proj$X, listings_sf_proj$longitude)

listings_sf_proj <- listings_sf_proj %>% 
  select(-latitude, -longitude)


listings_room <- listings_sf_proj %>% 
  filter(room_type == "Private room") %>% 
  sample_n(500) # takes long time!

listings_lisa <- lisa(listings_room$X, listings_room$Y, listings_room$price, 
                      neigh = 10, resamp = 1000, latlon = FALSE, quiet = FALSE)

plot(listings_lisa)

```

## `spdep`

```{r}
listings <- read_csv("./data/listings.csv") %>% 
  filter(room_type == "Private room") %>%
  # filter(room_type == "Entire home/apt") %>%
  filter(price > 1)

coordinates(listings) <- ~ longitude + latitude

View(listings)
class(listings)
str(listings)
names(listings)
head(listings@data)
head(listings@coords)

neigh_4 <- knn2nb(knearneigh(listings, k = 4, longlat = TRUE))
neigh_7 <- knn2nb(knearneigh(listings, k = 7, longlat = TRUE))
neigh_9 <- knn2nb(knearneigh(listings, k = 9, longlat = TRUE))

neigh_7
class(neigh_7)
names(neigh_7)

# plot(outline, border = "grey")
# plot(neigh_7, coordinates(listings))
# plot(diffnb(neigh_9, neigh_4), xy, add=TRUE, col="red", lty=2)
# title(main="Differences between Euclidean and Great Circle k=4 neighbours")

moran.test(listings$price, listw = nb2listw(neigh_7))

moran.mc(listings$price, listw = nb2listw(neigh_7), nsim = 999)

moran.plot(listings$price, listw = nb2listw(neigh_7))

lm <- localmoran(listings$price, listw = nb2listw(neigh_7))

summary(lm)

# standarize variables
listings$price_std <- scale(listings$price) 

# create a lagged variable
listings$price_lag <- lag.listw(nb2listw(neigh_7), listings$price_std)

# extract p value from local Moran
table(lm[, 5] <= 0.05)
listings$price_sig <- lm[, 5] <= 0.05

listings$quad_sig <- NA
listings@data[(listings$price_std > 0 & listings$price_lag > 0) & (lm[, 5] <= 0.05), "quad_sig"] <- 1
listings@data[(listings$price_std < 0 & listings$price_lag < 0) & (lm[, 5] <= 0.05), "quad_sig"] <- 2
listings@data[(listings$price_std > 0 & listings$price_lag < 0) & (lm[, 5] <= 0.05), "quad_sig"] <- 3
listings@data[(listings$price_std < 0 & listings$price_lag > 0) & (lm[, 5] <= 0.05), "quad_sig"] <- 4
listings@data[lm[, 5] >= 0.05, "quad_sig"] <- 5  

table(listings$quad_sig)

# Set the breaks for the thematic map classes
breaks <- seq(1, 5, 1)

# Set the corresponding labels for the thematic map classes
labels <- c("High-High", "Low-Low", "High-Low", "Low-High", "Not Signif.")

# see ?findInterval - This is necessary for making a map
np <- findInterval(listings$quad_sig, breaks)

# Assign lisa colors to each map class
tmap_mode("plot")

tm_shape(listings) +
  tm_dots(col = "quad_sig", style = "cat", 
          # palette = c("red", "blue", "lightpink", "skyblue2", "white"), 
          # labels = c("High-High", "Low-Low", "High-Low", "Low-High", "Not Signif."), 
          palette = c("red", "grey"), 
          labels = c("High-High", "Not Signif."), 
          title = "Local Moran's I")

tmap_mode("view")

st_as_sf(listings) %>% 
  filter(!is.na(quad_sig)) %>% 
  tm_shape() +
  tm_dots(col = "quad_sig", style = "cat", 
          palette = c("red", "blue", "lightpink", "skyblue2", "white"), 
          labels = c("High-High", "Low-Low", "High-Low", "Low-High", "Not Signif."), 
          title = "Local Moran's I",
          popup.vars = c("name", "price", "quad_sig"))

listings <- st_as_sf(listings) %>% 
  st_set_crs(4326) %>% 
  st_transform(3112)

temp <- st_join(listings, SA2_SEIFA_proj, join = st_intersects)

temp %>% 
  st_drop_geometry() %>% 
  filter(!is.na(IRSAD_d)) %>% 
  ggplot(aes(as.factor(quad_sig), price)) + 
  geom_boxplot() + scale_y_log10()

temp %>% 
  st_drop_geometry() %>% 
  filter(!is.na(IRSAD_d)) %>% 
  ggplot(aes(as.factor(IRSAD_d), fill = as.factor(quad_sig))) +
  geom_bar(position = "fill") +
  scale_y_continuous(labels = scales::percent)

```

# Correlogram 

http://www.petrkeil.com/?p=1050

```{r}
# method "corr" for correlation, "I" for Moran's I, "C" for Geary's C

listings_corr_corr <- sp.correlogram(neighbours = neigh_9, var = listings$price,
                                     order = 9, method = "corr", zero.policy = TRUE)
listings_corr_corr
plot(listings_corr_corr)

listings_corr_I <- sp.correlogram(neighbours = neigh_9, var = listings$price,
                                  order = 9, method = "I", zero.policy = TRUE)
listings_corr_I
plot(listings_corr_I)

listings_corr_C <- sp.correlogram(neighbours = neigh_9, var = listings$price,
                                  order = 9, method = "C", zero.policy = TRUE)
listings_corr_C
plot(listings_corr_C)

# library(pgirmess)
# listings_corr <- correlog(coordinates(listings), listings$price, method = "Moran")
# listings_corr
# plot(listings_corr)

```
