==> docs/99_resources.Rmd <==
---
title: "R and spatial data: resources"
author: "Radoslaw Panczak"
date: 2019-02-25
output: 
  html_document: 
    highlight: pygments
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
    toc_float: yes
---

# R

## Get help from humans

Brisbane Users of R Group - BURGr https://www.meetup.com/en-AU/Brisbane-Users-of-R-Group-BURGr/

Hacky Hour UQ - https://rcc.uq.edu.au/meetups#hacky-hour


## General 

Excellent (and freely available!) *R for Data Science* book comes very handy as a general R resource. Although it doesn't cover spatial data it will guide you through best practices of manipulating and graphing your data. Some chapters and parts  related to our work that can be useful are:

  - setting up and working with RStudio projects  https://r4ds.had.co.nz/workflow-projects.html
  - working with pipes https://r4ds.had.co.nz/transform.html#combining-multiple-operations-with-the-pipe
  - joining data https://r4ds.had.co.nz/relational-data.html

Fundamentals of Data Visualization book is fantastic resource to learn more about visualization (not only spatial data) https://serialmentor.com/dataviz/

## Spatial 

"Applied Spatial Data Analysis with R" book is a goto resouce explaining many concepts of spatial analysis and providing R code https://asdar-book.org/.

"An Introduction to R for Spatial Analysis and Mapping" book is an excellent resouce with a very applied focus https://uk.sagepub.com/en-gb/eur/an-introduction-to-r-for-spatial-analysis-and-mapping/book258267 (code available at https://study.sagepub.com/brunsdon2e)

Similarly "Geocomputation with R" offers both theoretical and practical approach to spatial data analysis https://bookdown.org/robinlovelace/geocompr/ 

"Spatial Data Science" is a book (in progress) by the author of `sf` package. Once finished that will be probably the authoritative text for this package https://www.r-spatial.org/book

"Intro to GIS and Spatial Analysis" is a free online book covering spatial R https://mgimond.github.io/Spatial/index.html

"Displaying Time Series, Spatial, and Space-Time Data with R" has heavier focus on visualizationa and also incorporates temporal dimension https://www.crcpress.com/Displaying-Time-Series-Spatial-and-Space-Time-Data-with-R-Second-Edition/Lamigueiro/p/book/9781138089983 (book website https://oscarperpinan.github.io/bookvis/)

Data Carpentry’s Geospatial Workshop http://datacarpentry.org/geospatial-workshop/

Learn more about simple features and `sf` library https://r-spatial.github.io/sf/

Handy Simple Features (sf) cheatsheet https://github.com/rstudio/cheatsheets/raw/master/sf.pdf 

Learn more about `tmap` library https://github.com/mtennekes/tmap


# Geography 

## Projections

Short video introducing projections https://www.youtube.com/watch?v=kIID5FDi2JQ 

Myriahedral Projections: Jarke J. van Wijk. Unfolding the Earth: Myriahedral Projections. *The Cartographic Journa*l, Vol. 45, No. 1, pp.32-42, February 2008. https://www.win.tue.nl/~vanwijk/myriahedral/ 


## Cartography

A short, friendly guide to basic principles of map design by axismaps https://www.axismaps.com/guide/

Which map should you choose? Visual Vocabulary for spatial data https://gramener.github.io/visual-vocabulary-vega/#/Spatial/

"How to Lie with Maps" book will teach you to recognize (and avoid bad) tricks involved in map making https://books.google.com.au/books/about/How_to_Lie_with_Maps.html?id=Q8OHiOiYwYUC&redir_esc=y (book review http://www.cartographicperspectives.org/index.php/journal/article/view/1498/1660)

## Geospatial analysis

Theory behind many ideas and concepts can be found in "Geospatial Analysis - A comprehensive guide - 2018" book available at https://www.spatialanalysisonline.com/ (browse content at https://www.spatialanalysisonline.com/HTML/index.html). Chapter on global autocorrelation for instance is here https://www.spatialanalysisonline.com/HTML/index.html?global_spatial_autocorrelation.htm and local - here https://www.spatialanalysisonline.com/HTML/index.html?local_indicators_of_spatial_as.htm

Original work on Moran's I is: Moran, P.A.P. (1950). "Notes on Continuous Stochastic Phenomena". *Biometrika*. **37**(1): 17–23. https://doi.org/10.2307/2332142

Original work on LISA is: Anselin, L. (1995)  "Local indicators of spatial association - LISA". *Geographical Analysis* **27**:93-115. https://doi.org/10.1111/j.1538-4632.1995.tb00338.x


# Inspiration

Some pretty R based spatial visualizations http://spatial.ly/r/

Reddit gallery of (often unusual, sometimes amusing) maps https://www.reddit.com/r/MapPorn/
